+++ zcash_client_backend/CHANGELOG.md
@@ -55,6 +55,7 @@ and this library adheres to Rust's notion of
     - `fn with_orchard_tree_mut`
     - `fn put_orchard_subtree_roots`
   - Added method `WalletRead::validate_seed`
+  - Removed `Error::AccountNotFound` variant.
 - `zcash_client_backend::fees`:
   - Arguments to `ChangeStrategy::compute_balance` have changed.